# [KMSv2] Add tracing