- Add tracing for KMSv2 encrypt/decrypt operations

fixes https://github.com/kubernetes/kubernetes/issues/120881

/kind feature
/sig auth
/triage accepted
/milestone v1.29
/priority important-soon

```release-note
Enable traces for KMSv2 encrypt/decrypt operations.
```

```docs
[KEP] https://github.com/kubernetes/enhancements/tree/master/keps/sig-auth/3299-kms-v2-improvements
```