@enj 2023-10-20
@aramase something questions:

Should we enhance

https://github.com/kubernetes/kubernetes/blob/ad698cc0aeb6e9b5fb1a7913d5db9f65fb1a51ac/staging/src/k8s.io/apiserver/pkg/storage/value/encrypt/envelope/kmsv2/grpc_service.go#L145-L153

to include tracing as well?

How hard would it to be add an integration test?

Do you have an image that shows how these spans will look?

/milestone v1.29