@aramase 2023-10-23
https://github.com/kubernetes/kubernetes/blob/ad698cc0aeb6e9b5fb1a7913d5db9f65fb1a51ac/staging/src/k8s.io/apiserver/pkg/storage/value/encrypt/envelope/kmsv2/grpc_service.go#L145-L153

Given the grpc call is only made once during start up/rotation and isn't part of every request, I decided not to add the tracing here. Let me know what you think!

> How hard would it to be add an integration test?

Added an integration test.

> Do you have an image that shows how these spans will look?

Haven't been able to successfully visualize the API server traces with Jaeger but the integration tests are validating it.

/cc dashpole