@aramase 2023-10-23
> I am deferring `/lgtm` to `sig-instrumentation-reviewers`

/assign @dashpole