@aramase 2023-10-24
> As the issue is tagged for 1.29, is it still planned for that release?

@moficodes Yes, this is still planned for `v1.29`. This PR is already approved. I've tagged @dashpole for review + lgtm.