@aramase 2023-10-26
Running existing benchmarks before and after tracing

Commit before tracing was added:
git checkout c2a04fa1cf5f39ddcab1b7591d06292c0b8487fb

```bash
go test ./test/integration/controlplane/transformation -bench BenchmarkKMSv2KDF -benchmem -run X -count=15 > kmsv2kdf_without_tracing.txt
go test ./test/integration/controlplane/transformation -bench BenchmarkKMSv2REST -benchmem -run X -count=20 > kmsv2_without_tracing.txt
```

Commit tracing was added
git checkout [8d3a25c7c98d77419111a02917f459aab8970087](https://github.com/kubernetes/kubernetes/commit/8d3a25c7c98d77419111a02917f459aab8970087)

```bash
go test ./test/integration/controlplane/transformation -bench BenchmarkKMSv2KDF -benchmem -run X -count=15 > kmsv2kdf_with_tracing.txt
go test ./test/integration/controlplane/transformation -bench BenchmarkKMSv2REST -benchmem -run X -count=20 > kmsv2_with_tracing.txt
```

```
benchstat kmsv2kdf_without_tracing.txt kmsv2kdf_with_tracing.txt

goos: linux
goarch: amd64
pkg: k8s.io/kubernetes/test/integration/controlplane/transformation
cpu: AMD EPYC 7763 64-Core Processor
           │ kmsv2kdf_without_tracing.txt │   kmsv2kdf_with_tracing.txt    │
           │            sec/op            │   sec/op     vs base           │
KMSv2KDF-8                    24.58 ± ∞ ¹   17.56 ± ∞ ¹  ~ (p=1.000 n=1) ²
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05

           │ kmsv2kdf_without_tracing.txt │    kmsv2kdf_with_tracing.txt     │
           │             B/op             │     B/op       vs base           │
KMSv2KDF-8                  16.02Gi ± ∞ ¹   16.02Gi ± ∞ ¹  ~ (p=1.000 n=1) ²
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05

           │ kmsv2kdf_without_tracing.txt │    kmsv2kdf_with_tracing.txt    │
           │          allocs/op           │  allocs/op    vs base           │
KMSv2KDF-8                   1.124M ± ∞ ¹   1.193M ± ∞ ¹  ~ (p=1.000 n=1) ²
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05

benchstat kmsv2_without_tracing.txt kmsv2_with_tracing.txt
goos: linux
goarch: amd64
pkg: k8s.io/kubernetes/test/integration/controlplane/transformation
cpu: AMD EPYC 7763 64-Core Processor
            │ kmsv2_without_tracing.txt │     kmsv2_with_tracing.txt     │
            │          sec/op           │   sec/op     vs base           │
KMSv2REST-8                 22.72 ± ∞ ¹   21.93 ± ∞ ¹  ~ (p=1.000 n=1) ²
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05

            │ kmsv2_without_tracing.txt │      kmsv2_with_tracing.txt      │
            │           B/op            │     B/op       vs base           │
KMSv2REST-8               25.84Gi ± ∞ ¹   25.84Gi ± ∞ ¹  ~ (p=1.000 n=1) ²
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05

            │ kmsv2_without_tracing.txt │     kmsv2_with_tracing.txt      │
            │         allocs/op         │  allocs/op    vs base           │
KMSv2REST-8                3.361M ± ∞ ¹   3.437M ± ∞ ¹  ~ (p=1.000 n=1) ²
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05
```